added test for /usr/bin/local
apparently under certain conditions this directory may not exists on
macos. looked online to why, and by the looks of it, this dir is created
when certain homebrew installs requires it to be there.

now given i didn't have this dir already there, its possible for someone
else to have same dir structure as me resulting in installation failure.

deleted previous pr... well it got deleted on its own after i synced my
fork with upstream but ehh doesnt' matter. one of the later commits
cleaned the install script which caused the merge conflict. this commit
fixes that.